backend/pyproject.toml
[project]
name = "ai-game-sandbox"
version = "0.1.0"
description = "AI Video Game Asset Generator with FAL AI integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "anthropic",
    "loguru",
    "python-dotenv>=1.0.0",
    "pydantic==2.5.3",
    "fal-client>=0.4.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
]

[project.scripts]
generate-image = "image_generation.cli:main"
process-sprites = "sprite_processing.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["image_generation", "sprite_processing", "scene_builder"]

[tool.hatch.build]
include = [
    "image_generation/**/*.py",
    "sprite_processing/**/*.py",
    "scene_builder/**/*.py",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
